requirements.txt
certifi==2026.2.25
charset-normalizer==3.4.6
greenlet==3.3.2
idna==3.11
pillow==12.1.1
playwright==1.58.0
pyee==13.0.1
python-dotenv==1.2.2
reportlab==4.4.10
requests==2.32.5
typing_extensions==4.15.0
urllib3==2.6.3
